Database/database.py
import mysql.connector  # Import mysql.connector to connect to the MySQL database

def get_db_connection():
    # Function to establish a connection to the MySQL database
    conn = mysql.connector.connect(
        host='localhost',
        user='coeus_user',
        password='secure_password',
        database='CoeusAI'
    )
    return conn

def setup_database():
    # Function to set up the database (this is a placeholder as tables should be created beforehand)
    conn = get_db_connection()  # Establish a database connection
    cursor = conn.cursor()
    # Here you can add SQL commands to set up tables if needed
    conn.close()  # Close the connection

