Database/requirements.txt
mysql-connector-python
tkinter
